pyproject.toml
[tool.poetry]
name = "university-recommendation"
version = "0.1.0"
description = ""
authors = ["Yash Nagraj <pes1202203845@pesu.pes.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.10"
pydantic = "^2.9.2"
fastapi = "^0.115.4"
openai = "^1.53.0"
uvicorn = "^0.32.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
